* server-src/holding.c: Check return value of fgets.
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/branches/3_3@5344 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/holding.c
@@ -736,19 +736,20 @@ holding_cleanup_dir(
     if (pid_FILE) {
 	char line[1000];
 	int  pid;
-	fgets(line, 1000, pid_FILE);
-	pid = atoi(line);
-	if (pid != getpid()) {
-	    /* check if pid is alive */
-	    if (kill(pid, 0) == 0) {
-		if (data->verbose_output)
-		    g_fprintf(data->verbose_output,
-			_("..skipping running directory '%s'\n"), element);
-		g_free(pid_file);
-		return 0;
+	if (fgets(line, 1000, pid_FILE) > 0) {
+	    pid = atoi(line);
+	    if (pid != getpid()) {
+		/* check if pid is alive */
+		if (kill(pid, 0) == 0) {
+		    if (data->verbose_output)
+			g_fprintf(data->verbose_output,
+			    _("..skipping running directory '%s'\n"), element);
+		    g_free(pid_file);
+		    return 0;
+		}
 	    }
+	    unlink(pid_file);
 	}
-	unlink(pid_file);
     }
     g_free(pid_file);
